move thread context to a dedicated opentelemetry module
